+++ src/atmcirclib/icon.py
@@ -142,3 +142,46 @@ def convert_icon_time(time: xr.DataArray) -> xr.DataArray:
     return xr.DataArray(
         data=list(map(convert_icon_time_step, time)), coords={"time": time}
     )
+
+
+def convert_xr_time(time_: xr.DataArray) -> list[dt.datetime]:
+    """Convert an xarray time array to a list of datetime objects."""
+    return [ts.astype("datetime64[s]").astype(dt.datetime) for ts in time_.data]
+
+
+def disaggr_tot_precip(tot_prec: xr.DataArray) -> xr.DataArray:
+    """Disaggregate total precipitation into 'instantaneous' precipitation."""
+    prec = tot_prec.copy()
+    prec.data[1:] = prec.data[1:] - prec.data[:-1]
+    prec.attrs["units"] = deriv_precip_intensity_units(tot_prec)
+    return prec
+
+
+def deriv_precip_intensity_units(tot_prec: xr.DataArray, time_unit: str = "s") -> str:
+    """Derive the units of precipitation intensity from total precipitation.
+
+    Example: For ICON output written with a frequency of 10 min containing
+    total precipitation ('tot_prec'), 'mm / 600 s' is returned.
+
+    """
+    if (n := tot_prec.attrs["standard_name"]) not in (ns := ["tot_prec"]):
+        cs = ", ".join([f"'{n}'" for n in ns])
+        raise ValueError(f"unexpected standard name (choices: {cs}): '{n}'")
+    if (u := tot_prec.attrs["units"]) not in (us := ["kg m-2"]):
+        cs = ", ".join([f"'{u}'" for u in us])
+        raise ValueError(f"unexpected standard name (choices: {cs}): '{u}'")
+    if (d := "time") not in (ds := tot_prec.dims):
+        raise ValueError(f"unexpected dimensions (missing '{d}'): {ds}")
+    time_abs: xr.DataArray = convert_icon_time(tot_prec.time)
+    steps: list[dt.datetime] = convert_xr_time(time_abs)
+    deltas = [steps[i + 1] - t for i, t in enumerate(steps[:-1])]
+    delta = deltas[0]
+    if not all(d == delta for d in deltas):
+        raise Exception(f"deltas are no all the same: {deltas}")
+    units = "mm / "
+    secs = delta.total_seconds()
+    if time_unit == "s":
+        units += f"{secs:g} s"
+    else:
+        raise NotImplementedError(f"time unit other than 's': '{time_unit}'")
+    return units